requirements.txt
av
einops
flashy>=0.0.1
hydra-core>=1.1
hydra_colorlog
julius
num2words
numpy
sentencepiece
spacy==3.5.2
torch>=2.0.0
torchaudio>=2.0.0
huggingface_hub
tqdm
transformers>=4.31.0  # need Encodec there.
xformers
demucs
librosa
torchmetrics
encodec
streamlit
google-generativeai
python-dotenv